requirements.txt
etherscan-python
web3
pandas
datatable
